API, RestaurantList Działa
"Thus do we invoke the Machine God.
Thus do we make whole that which was sundered."~Hymn of Reforging

+++ app/src/main/java/com/leshen/letseatmobile/restaurantList/RestaurantListModel.kt
@@ -8,6 +8,6 @@ data class RestaurantListModel(
     val restaurantCategory: String,
     val restaurantId: Int,
     val restaurantName: String,
-    val stars: Int,
+    val stars: Double,
     val tableModels: List<TableModel>
 )
